Changed Dispatch_Count_Handler::timers_fired_ from int to size_t to make a good match with ACE_MAX_TIMERS.

+++ tests/MT_Reactor_Timer_Test.h
@@ -102,7 +102,7 @@ private:
 
   int input_seen_;
   int notify_seen_;
-  int timers_fired_;
+  size_t timers_fired_;
 
 };
